firebese adm

+++ lib/models/product.dart
@@ -1,26 +1,32 @@
 import 'dart:io';
-
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_storage/firebase_storage.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:lojavirtual/models/item_size.dart';
 import 'package:uuid/uuid.dart';
 
-/// MODELO PARA SALVAR NO FIREBASE
-
 class Product extends ChangeNotifier {
-  Product({this.id, this.name, this.description, this.images, this.sizes}) {
+
+  Product({this.id, this.name, this.description, this.images, this.sizes,
+    this.deleted = false}){
     images = images ?? [];
     sizes = sizes ?? [];
+  }
 
-    /// INICIALIZA COMO PRODUTOS VAZIOS PARA PODER CRIAR UM NOVO PRODUTO, AGORA Ñ SÃO MAIS NULOS.
+  Product.fromDocument(DocumentSnapshot document){
+    id = document.id;
+    name = document['name'] as String;
+    description = document['description'] as String;
+    images = List<String>.from(document.data()['images'] as List<dynamic>);
+    deleted = (document.data()['deleted'] ?? false) as bool;
+    sizes = (document.data()['sizes'] as List<dynamic> ?? []).map(
+            (s) => ItemSize.fromMap(s as Map<String, dynamic>)).toList();
   }
 
   final FirebaseFirestore firestore = FirebaseFirestore.instance;
   final FirebaseStorage storage = FirebaseStorage.instance;
 
   DocumentReference get firestoreRef => firestore.doc('products/$id');
-
   Reference get storageRef => storage.ref().child('products').child(id);
 
   String id;
@@ -28,69 +34,70 @@ class Product extends ChangeNotifier {
   String description;
   List<String> images;
   List<ItemSize> sizes;
-  ItemSize _selectedSize;
 
   List<dynamic> newImages;
 
-  Product.fromDocument(DocumentSnapshot document) {
-    id = document.id;
-    name = document['name'] as String;
-    description = document['description'] as String;
-    images = List<String>.from(document.data()['images'] as List<dynamic>);
-    sizes = (document.data()['sizes'] as List<dynamic> ?? [])
-        .map((s) => ItemSize.fromMap(s as Map<String, dynamic>))
-        .toList();
+  bool deleted;
+
+  bool _loading = false;
+  bool get loading => _loading;
+  set loading(bool value){
+    _loading = value;
+    notifyListeners();
   }
 
+  ItemSize _selectedSize;
   ItemSize get selectedSize => _selectedSize;
+  set selectedSize(ItemSize value){
+    _selectedSize = value;
+    notifyListeners();
+  }
 
   int get totalStock {
     int stock = 0;
-    for (final size in sizes) {
+    for(final size in sizes){
       stock += size.stock;
     }
     return stock;
   }
 
   bool get hasStock {
-    return totalStock > 0;
+    return totalStock > 0 && !deleted;
   }
 
   num get basePrice {
     num lowest = double.infinity;
-    for (final size in sizes) {
-      if (size.price < lowest && size.hasStock) {
+    for(final size in sizes){
+      if(size.price < lowest) {
         lowest = size.price;
       }
     }
     return lowest;
   }
 
-  set selectedSize(ItemSize value) {
-    _selectedSize = value;
-    notifyListeners();
-  }
-
-  ItemSize findSize(String name) {
+  ItemSize findSize(String name){
     try {
       return sizes.firstWhere((s) => s.name == name);
-    } catch (e) {
+    } catch (e){
       return null;
     }
   }
 
-  List<Map<String, dynamic>> exportSizeList() {
+  List<Map<String, dynamic>> exportSizeList(){
     return sizes.map((size) => size.toMap()).toList();
   }
 
   Future<void> save() async {
+    loading = true;
+
     final Map<String, dynamic> data = {
       'name': name,
       'description': description,
       'sizes': exportSizeList(),
+      'deleted': deleted
     };
 
-    if (id == null) {
+    if(id == null){
       final doc = await firestore.collection('products').add(data);
       id = doc.id;
     } else {
@@ -99,46 +106,52 @@ class Product extends ChangeNotifier {
 
     final List<String> updateImages = [];
 
-    for (final newImage in newImages) {
-      if (images.contains(newImage)) {
+    for(final newImage in newImages){
+      if(images.contains(newImage)){
         updateImages.add(newImage as String);
       } else {
-        final UploadTask task =
-            storageRef.child(Uuid().v1()).putFile(newImage as File);
+        final UploadTask task = storageRef.child(const Uuid().v1()).putFile(newImage as File);
         final TaskSnapshot snapshot = await task;
         final String url = await snapshot.ref.getDownloadURL();
         updateImages.add(url);
       }
     }
 
-    for (final image in images) {
-      if (!newImages.contains(image)) {
+    for(final image in images){
+      if(!newImages.contains(image) && image.contains('firebase')){
         try {
-          final ref = storage.refFromURL(image);
+          final ref = storage.ref(image);
           await ref.delete();
-        } catch (e) {
-          debugPrint('falha ao deletar $image');
+        } catch (e){
+          debugPrint('Falha ao deletar $image');
         }
       }
     }
-    
+
     await firestoreRef.update({'images': updateImages});
+
+    images = updateImages;
+
+    loading = false;
   }
 
-  Product clone() {
+  void delete(){
+    firestoreRef.update({'deleted': true});
+  }
+
+  Product clone(){
     return Product(
       id: id,
       name: name,
       description: description,
       images: List.from(images),
       sizes: sizes.map((size) => size.clone()).toList(),
+      deleted: deleted,
     );
-
-    ///clone do original
   }
 
   @override
   String toString() {
-    return 'Product{id: $id, name: $name, description: $description, images: $images, sizes: $sizes}';
+    return 'Product{id: $id, name: $name, description: $description, images: $images, sizes: $sizes, newImages: $newImages}';
   }
 }
